refactor(flowcontrol): move the priority queue out of the plugins tree (#2146)
With the queue plugin registry removed (#2036), the priority queue is a
plain internal component of flow control, not a plugin. Move it from
pkg/epp/flowcontrol/framework/plugins/queue to pkg/epp/flowcontrol/queue,
deleting the now-empty pkg/epp/flowcontrol/framework tree so the only
"framework" package left is the plugin SDK at pkg/epp/framework.

Update the flow-control README pillars that still described the deleted
sub-package and the long-removed QueueCapability system.
